Add more tests.
